+++ lib/src/tags/builder.dart
@@ -41,35 +41,49 @@ class BuilderTag implements Tag {
   String get name => "builder";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies
+  ) {
     // 'vars' max length is 5
     final vars = parseListOfStrings(element.getAttribute("vars"));
-    if (vars != null && vars.length > 5) throw Exception("<$name> 'vars' attribute only accepts up to 5 variables");
+    if (vars != null && vars.length > 5){
+      throw Exception("<$name> 'vars' attribute only accepts up to 5 variables");
+    }
 
     // 'for' is a required attribute.
     final forAttribute = element.getAttribute("for");
-    if (forAttribute == null || forAttribute.isEmpty) throw Exception("<$name> 'for' attribute is required.");
+    if (forAttribute == null || forAttribute.isEmpty) {
+      throw Exception("<$name> 'for' attribute is required.");
+    }
 
     final multiChild = parseBool(attributes["multiChild"]) ?? false;
     final nullable = parseBool(attributes["nullable"]) ?? false;
     final dependenciesScope = attributes["dependenciesScope"];
 
     dynamic builder([p0, p1, p2, p3, p4]) {
-      final deps = XWidget.scopeDependencies(dependencies, dependenciesScope);
+      final params = <String, dynamic>{};
       if (vars != null) {
         for (int paramIndex = 0; paramIndex < vars.length; paramIndex++) {
           final varName = vars[paramIndex];
-          if (varName.isNotEmpty && !_ignoreVar.hasMatch(varName)) {
+          if (CommonUtils.isNotBlank(varName) && !_ignoreVar.hasMatch(varName)) {
             switch (paramIndex) {
-              case 0: if (p0 is! BuildContext) deps[varName] = p0; break;
-              case 1: if (p1 is! BuildContext) deps[varName] = p1; break;
-              case 2: if (p2 is! BuildContext) deps[varName] = p2; break;
-              case 3: if (p3 is! BuildContext) deps[varName] = p3; break;
-              case 4: if (p4 is! BuildContext) deps[varName] = p4; break;
+              case 0: if (p0 is! BuildContext) params[varName] = p0; break;
+              case 1: if (p1 is! BuildContext) params[varName] = p1; break;
+              case 2: if (p2 is! BuildContext) params[varName] = p2; break;
+              case 3: if (p3 is! BuildContext) params[varName] = p3; break;
+              case 4: if (p4 is! BuildContext) params[varName] = p4; break;
             }
           }
         }
       }
+      final deps = XWidget.scopeDependencies(
+          element,
+          dependencies,
+          dependenciesScope,
+          params.isEmpty ? "inherit" : "copy"
+      ).addAll(params);
       final children = XWidget.inflateXmlElementChildren(element, deps).objects;
       return multiChild ? children : XWidgetUtils.getOnlyChild("<$name> tag", children);
     }