+++ lib/src/tags/callback.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:xml/xml.dart';
 
 import '../utils/parsers.dart';
+import '../utils/utils.dart';
 import '../xwidget.dart';
 
 /// This tag allows you to bind an event handler with custom arguments. If you don't need to pass any
@@ -28,18 +29,28 @@ class CallbackTag implements Tag {
   String get name => "callback";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies
+  ) {
     // 'for' is a required attribute.
     final forAttribute = element.getAttribute("for");
-    if (forAttribute == null || forAttribute.isEmpty) throw Exception("<$name> 'for' attribute is required.");
+    if (forAttribute == null || forAttribute.isEmpty) {
+      throw Exception("<$name> 'for' attribute is required.");
+    }
 
     // 'vars' max length is 5
     final vars = parseListOfStrings(element.getAttribute("vars"));
-    if (vars != null && vars.length > 5) throw Exception("<$name> 'vars' attribute only accepts up to 5 variables");
+    if (vars != null && vars.length > 5) {
+      throw Exception("<$name> 'vars' attribute only accepts up to 5 variables");
+    }
 
     // 'action' is a required attribute
     final action = element.getAttribute("action");
-    if (action == null || action.isEmpty) throw Exception("<$name> 'action' attribute is required");
+    if (action == null || action.isEmpty) {
+      throw Exception("<$name> 'action' attribute is required");
+    }
 
     // 'returnVar' is optional
     final returnVar = element.getAttribute("returnVar");
@@ -48,23 +59,30 @@ class CallbackTag implements Tag {
     final dependenciesScope = attributes["dependenciesScope"];
 
     dynamic callback([p0, p1, p2, p3, p4]) {
-      final deps = XWidget.scopeDependencies(dependencies, dependenciesScope);
+      final params = <String, dynamic>{};
       if (vars != null) {
         for (int paramIndex = 0; paramIndex < vars.length; paramIndex++) {
           final varName = vars[paramIndex];
-          if (varName.isNotEmpty && !_ignoreVar.hasMatch(varName)) {
+          if (CommonUtils.isNotBlank(varName) && !_ignoreVar.hasMatch(varName)) {
             switch (paramIndex) {
-              case 0: if (p0 is! BuildContext) deps[varName] = p0; break;
-              case 1: if (p1 is! BuildContext) deps[varName] = p1; break;
-              case 2: if (p2 is! BuildContext) deps[varName] = p2; break;
-              case 3: if (p3 is! BuildContext) deps[varName] = p3; break;
-              case 4: if (p4 is! BuildContext) deps[varName] = p4; break;
+              case 0: if (p0 is! BuildContext) params[varName] = p0; break;
+              case 1: if (p1 is! BuildContext) params[varName] = p1; break;
+              case 2: if (p2 is! BuildContext) params[varName] = p2; break;
+              case 3: if (p3 is! BuildContext) params[varName] = p3; break;
+              case 4: if (p4 is! BuildContext) params[varName] = p4; break;
             }
           }
         }
       }
-      final value = XWidget.parseExpression(action, dependencies);
-      if (returnVar != null && returnVar.isNotEmpty) {
+      final hasReturnVar = returnVar != null && returnVar.isNotEmpty;
+      final deps = XWidget.scopeDependencies(
+          element,
+          dependencies,
+          dependenciesScope,
+          params.isEmpty && !hasReturnVar ? "inherit" : "copy"
+      ).addAll(params);
+      final value = XWidget.parseExpression(action, deps);
+      if (hasReturnVar) {
         deps.setValue(returnVar, value);
       }
     }