+++ lib/src/tags/fragment.dart
@@ -4,13 +4,16 @@ import '../xwidget.dart';
 
 /// A tag that renders a UI fragment
 ///
-/// Supports folders, inherited attributes, and HTML style query parameters. Parameters are stored as dependencies.
+/// Supports folders, inherited attributes, and HTML style query parameters.
+/// Parameters are stored as dependencies.
 ///
 /// Attributes:
-/// - name (required): name of the fragment to render i.e 'login' or 'login.xml'. You can prepend a path if
-///         if you're using fragment folders i.e. 'profile/login'.
-/// - dependenciesScope (optional): Defines the method for passing Dependencies to immediate children. Valid
-///         values are `new`, `copy`, and `inherit`. The default is `inherit`.
+/// - name (required): name of the fragment to render i.e 'login' or
+///        'login.xml'. You can prepend a path if you're using fragment
+///         folders i.e. 'profile/login'.
+/// - dependenciesScope (optional): Defines the method for passing Dependencies
+///         to immediate children. Valid values are `new`, `copy`, and
+///         `inherit`. The default is `inherit`.
 /// - for (optional): name of parent attribute to render the fragment into.
 ///         ```dart
 ///         <AppBar>
@@ -24,21 +27,42 @@ class FragmentTag implements Tag {
   String get name => "fragment";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies
+  ) {
     // 'name' is a required attribute.
-    final fragmentName = attributes["name"];
+    final String? fragmentName = attributes["name"];
     if (fragmentName == null) {
       final dump = XWidget.dump(element, dependencies);
       throw Exception("<$name> 'name' attribute is required. $dump");
     }
 
-    // 'dependenciesScope' is optional
-    final dependenciesScope = attributes["dependenciesScope"];
+    // process child tags for parameters only - need original deps here
+    // todo: only allow processing of 'params', 'forEach', and 'if' elements
+    final params = _onlyParams(XWidget.inflateXmlElementChildren(
+        element,
+        dependencies,
+        excludeText: true
+    ));
+
+    // scope dependencies. if we're going to modify the dependencies via params,
+    // then default the dependencies scope to 'copy'; otherwise default to
+    // 'inherit'.
+    final deps = XWidget.scopeDependencies(
+        element,
+        dependencies,
+        attributes["dependenciesScope"],
+        params.isNotEmpty || fragmentName.contains("?") ? "copy" : "inherit"
+    );
 
     // inflate named fragment
-    final deps = XWidget.scopeDependencies(dependencies, dependenciesScope);
-    final inheritedAttributes = element.attributes.where((attribute) => !attributeNames.contains(attribute.localName));
-    final object = XWidget.inflateFragment(fragmentName, deps, inheritedAttributes: inheritedAttributes);
+    final object = XWidget.inflateFragment(fragmentName, deps, params: params,
+        inheritedAttributes: element.attributes.where((attribute) {
+          return !attributeNames.contains(attribute.localName);
+        }),
+    );
     if (object == null) return null;
 
     final children = Children();
@@ -49,5 +73,19 @@ class FragmentTag implements Tag {
       children.objects.add(object);
     }
     return children;
+  }
+
+  //===================================
+  // private methods
+  //===================================
+
+  Map<String, dynamic> _onlyParams(Children children) {
+    final params = <String, dynamic>{};
+    for (final child in children.objects) {
+      if (child is MapEntry) {
+        params[child.key] = child.value;
+      }
+    }
+    return params;
   }
 }