+++ lib/src/tags/variable.dart
@@ -7,13 +7,22 @@ class VariableTag implements Tag {
   String get name => "var";
 
   @override
-  Children? processTag(XmlElement element, Map<String, dynamic> attributes, Dependencies dependencies) {
+  Children? processTag(
+      XmlElement element,
+      Map<String, dynamic> attributes,
+      Dependencies dependencies
+  ) {
     // 'name' is a required attribute
     final varName = attributes["name"];
-    if (varName == null || varName.isEmpty) throw Exception("<$name> 'name' attribute is required.");
+    if (varName == null || varName.isEmpty) {
+      throw Exception("<$name> 'name' attribute is required.");
+    }
 
-    // 'value' is a required attribute - check element attributes because the value could be null.
-    if (element.getAttribute("value") == null) throw Exception("<$name> 'value' attribute is required.");
+    // 'value' is a required attribute - check element attributes because
+    // the value could be null.
+    if (element.getAttribute("value") == null) {
+      throw Exception("<$name> 'value' attribute is required.");
+    }
 
     // set data value and exit
     final value = attributes["value"];
@@ -21,13 +30,12 @@ class VariableTag implements Tag {
       if (value is Map<String, dynamic>) {
         dependencies.addAll(value);
       } else {
-        throw Exception("Invalid type '${value.runtimeType}' for spread operator. "
-            "Expected 'Data' or 'Map<String, dynamic>'.");
+        throw Exception("Invalid type '${value.runtimeType}' for spread "
+            "operator. Expected 'Data' or 'Map<String, dynamic>'.");
       }
     } else {
       dependencies.setValue(varName, value);
     }
-
     return null;
   }
 }