+++ lib/xwidget.dart
@@ -1,8 +1,8 @@
 export 'src/el/parser.dart';
 
 export 'src/custom/async.dart';
+export 'src/custom/collection.dart';
 export 'src/custom/controller.dart';
-export 'src/custom/list_of.dart';
 export 'src/custom/value_listener.dart';
 
 export 'src/utils/brackets.dart';